% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency_exchange_rates.R
\name{get_currency_exchange_rates}
\alias{get_currency_exchange_rates}
\title{Get currency exchange rates for a given quarter from the IMF API}
\usage{
get_currency_exchange_rates(quarter, max_seconds = 30L)
}
\arguments{
\item{quarter}{A single string containing the desired quarter of data in the
form "2022-Q4".}

\item{max_seconds}{A single numeric containing the max number of seconds it
should wait for a response}
}
\value{
A dataframe containing columns \code{currency} and \code{exchange_rate}
}
\description{
Get currency exchange rates for a given quarter from the IMF API
}
