% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_regions.R
\name{get_index_regions}
\alias{get_index_regions}
\title{Get the definition of MSCI regions (in terms of member countries)}
\usage{
get_index_regions()
}
\value{
A dataframe containing columns \code{equity_market}, \code{country}, and
\code{country_iso}
}
\description{
Get the definition of MSCI regions (in terms of member countries)
}
