% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ishares_index_data.R
\name{get_ishares_index_data}
\alias{get_ishares_index_data}
\title{Scrape index data from the iShares website}
\usage{
get_ishares_index_data(url, name, as_of_date)
}
\arguments{
\item{url}{A string containing the url of the desired iShares index data.}

\item{name}{A string containing the name of the index.}

\item{as_of_date}{A string indicating the desired as_of_date, by year, month
and date, "YYYYMMDD" (e.g. "20201231"). Data will be scraped for date
specified.}
}
\value{
A data.frame, containing the un-processed iShares data.
}
\description{
This function will scrape and process holdings data from the iShares website.
}
\examples{
\dontrun{
url <-
  paste0(
    "https://www.ishares.com/uk/individual/en/products/",
    "251813/ishares-global-corporate-bond-ucits-etf/"
  )
name <- "iShares Global Corporate Bond UCITS ETF <USD (Distributing)>"
as_of_date <- "20211231"

get_ishares_index_data(url, name, as_of_date)
}

}
