% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ishares_index_data.R
\name{process_ishares_index_data}
\alias{process_ishares_index_data}
\title{Process raw iShares data into PACTA readable format}
\usage{
process_ishares_index_data(data)
}
\arguments{
\item{data}{An iShares index dataset, like the output of
\code{get_ishares_index_data()}.}
}
\value{
A processed dataset, formatted for input into
\code{pacta.portfolio.analysis}.
}
\description{
Process raw iShares data into PACTA readable format
}
